move `ordered_map` in the `io` module
